bomba vest adjustment

+++ code/game/objects/items/explosives/bombvest.dm
@@ -18,6 +18,8 @@
 	if(activator.wear_suit != src)
 		to_chat(activator, "Due to the rigging of this device, it can only be detonated while worn.") //If you are going to use this, you have to accept death. No armor allowed.
 		return FALSE
+	if(!do_after(user, 3, TRUE, src, BUSY_ICON_DANGER, ignore_turf_checks = TRUE))
+		return FALSE
 	if(bomb_message) //Checks for a non null bomb message.
 		activator.say("[bomb_message]!!")
 		message_admins("[activator] has detonated an explosive vest with the warcry \"[bomb_message]\".") //Incase disputes show up about marines killing themselves and others.
@@ -25,7 +27,8 @@
 	else
 		message_admins("[activator] has detonated an explosive vest with no warcry.")
 		log_game("[activator] has detonated an explosive vest with no warcry.")
-	explosion(loc, 0, 2, 6, 5, 5) 
+	activator.adjustBruteLoss(1200) //Oops we blew all our limbs off
+	explosion(loc, 0, 2, 6, 5, 5)
 	qdel(src)
 
 ///Gets a warcry to scream on Control Click, checks for non allowed warcries.
@@ -53,8 +56,11 @@
 	if(activator.wear_suit != src)
 		to_chat(activator, "Due to the rigging of this device, it can only be detonated while worn.")
 		return FALSE
+	if(!do_after(user, 3, TRUE, src, BUSY_ICON_DANGER, ignore_turf_checks = TRUE))
+		return FALSE
 	activator.say("ORBITAL BOMBARDMENT INBOUND!!")
 	message_admins("[activator] has detonated an Orbital Bombardment vest! Unga!")
 	log_game("[activator] has detonated an Orbital Bombardment vest! Unga!")
+	activator.adjustBruteLoss(1200) //Oops we blew all our limbs off
 	explosion(loc, 15, 15, 15, 15, 15)
 	qdel(src)
